feat(mapview): updates to set search area + Toast when results is empty + other updates

+++ PostosEletricos/Domain/GoogleMaps/GoogleMapsAPI.swift
@@ -9,7 +9,7 @@ import Foundation
 import Moya
 
 enum GoogleMapsAPI {
-    case eletricalChargingStations(latitude: Double, longitude: Double, radius: Float)
+    case eletricalChargingStations(latitude: Double, longitude: Double, radius: Double)
 }
 
 extension GoogleMapsAPI: TargetType {
